refactor(channelui): move messages and layout helpers into channelui
Hoists channel_messages.go and channel_layout.go from package main into
cmd/wuphf/channelui. PR 4b of the channel-package extraction series,
stacked on refactor/channelui-renderers.

Moved (with PascalCase exports):
- countReplies         → channelui.CountReplies
- buildReplyChildren   → channelui.BuildReplyChildren
- parseTimestamp       → channelui.ParseTimestamp
- formatShortTime      → channelui.FormatShortTime
- layoutDimensions     → channelui.LayoutDimensions
- computeLayout        → channelui.ComputeLayout
- renderVerticalBorder → channelui.RenderVerticalBorder

Function and type aliases in package main (channelui_aliases.go) keep
the lowercase names callable so no caller code changes:

    var (
        countReplies  = channelui.CountReplies
        ...
        computeLayout = channelui.ComputeLayout
    )
    type layoutDimensions = channelui.LayoutDimensions

Both helpers were leafmost — fully self-contained beyond BrokerMessage
(already in channelui) and the lipgloss/strings/time stdlib. No other
package-main symbol moves were needed.

No behavior change. 34 packages green, golangci clean, binary builds.
